+++ CHANGELOG.md
@@ -44,6 +44,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ### Fixed  
  - Routing for projects merged (`/conversion-project` and `/transfer-project` become `/project`)
  - "Team projects" > "Handed over" now shows unassigned projects again
+ - Authorization fixed on all API endpoints
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-08.217...production-2025-05-13.244) for everything in the release
 ---